Fix Makefile issue and formatting.
This change fixes an issue that crops up when pb2json's build is invoked
from within a CMake build process (e.g., as an external library target).
CMake gets confused by the tests being built alongside the library and
believes that the working directory isn't clean after the first build.

To fix this, tests are now a separate build target at the top level.

The change also improves the Makefile formatting and uses the ${CXX}
variable instead of hard-coding use of g++ as a compiler.